Factored out comparison data into its own store
